move the sign up form to independent component
