+++ src/containers/Home.js
@@ -1,10 +1,14 @@
-const simpleRequest = async () => {
+import { connect } from 'react-redux';
+import PropTypes from 'prop-types';
+
+const simpleRequest = async (authToken) => {
   try {
     const options = {
       method: 'GET',
       headers: {
         Accept: 'application/json',
         'Content-Type': 'application/json',
+        Authorization: authToken,
       },
     };
     const response = await fetch('http://[::1]:3000/cars', options);
@@ -15,12 +19,20 @@ const simpleRequest = async () => {
   }
 };
 
-const Home = () => {
-  simpleRequest();
+const Home = ({ authToken }) => {
+  simpleRequest(authToken);
   return (
     <div>
       <h1>Welcome Home</h1>
     </div>
   );
 };
-export default Home;
+
+Home.propTypes = {
+  authToken: PropTypes.string.isRequired,
+};
+
+const mapStateToProp = (state) => ({
+  authToken: state,
+});
+export default connect(mapStateToProp)(Home);
